requirements.txt
Django==1.3
MySQL-python==1.2.3
South==0.7.3

django-compressor==0.9.2
django-debug-toolbar==0.8.5
django-extensions==0.6
Fabric==1.0.1

BeautifulSoup==3.2.0
html5lib==0.90
markdown2==1.0.1.16
python-memcached==1.45
python-openid==2.2.5
recaptcha-client==1.0.6